refactor(profile): Refactor profile view tab logic in into its own mixin file
