+++ core/network/profiles/views.py
@@ -3,7 +3,6 @@
 
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.shortcuts import get_object_or_404, render
-from django.template.loader import render_to_string
 from django.urls import reverse_lazy
 from django.views.generic import (
     ListView,
@@ -14,69 +13,12 @@ from django.views.generic import (
 
 from network.posts.models import Post
 
-from .constants import photo_tabs, profile_tabs
 from .forms import ProfileForm
+from .mixins import PartialPhotoTabMixin, ProfileContextMixin, ProfileTabMixin
 from .models import Profile
 
 
-class TabMixin:
-    """Provide tab context data and handle partial & full tab page request."""
-
-    profile_tab = None
-    active_tab = None
-
-    def get_tab_context(self):
-        """Get tab context. Update photo tab context if it's photo tab."""
-        context = {
-            "tabs": profile_tabs,
-            "current_tab": self.profile_tab,
-        }
-        if self.profile_tab == "photos":
-            context.update(
-                {
-                    "photo_tabs": photo_tabs,
-                    "current_photo_tab": self.active_tab,
-                }
-            )
-        return context
-
-    def is_partial_request(self):
-        """Check if request is partial or full for profile tab request."""
-        return self.request.GET.get("is_partial_request", None) == "true"
-
-    def get_sub_tab_template_path(self):
-        """Dynamically return tab template sub path."""
-        request_page_str = self.request.path.rstrip("/").split("/")[-1]
-
-        return (
-            f"partial/{request_page_str}"
-            if self.is_partial_request()
-            else f"full/{request_page_str}"
-        )
-
-    def get_tab_template_path(self):
-        """Dynamically return full or partial template path."""
-        return f"profiles/tabs/{self.get_sub_tab_template_path()}.html"
-
-
-class ProfileContextMixin:
-    """Mixin to provide profile to context."""
-
-    def dispatch(self, request, *args, **kwargs):
-        """Set profile instance on receiving request."""
-        self.set_profile()
-        return super().dispatch(request, *args, **kwargs)
-
-    def set_profile(self):
-        """Get profile instance from argument or return 404."""
-        self._profile = get_object_or_404(Profile, username=self.kwargs["username"])
-
-    def get_profile_context(self):
-        """Return get_profile_context."""
-        return {"profile": self._profile}
-
-
-class ProfileBaseTabView(ProfileContextMixin, TabMixin, TemplateView):
+class ProfileBaseTabView(ProfileContextMixin, ProfileTabMixin, TemplateView):
     """Profile base view."""
 
     def get_template_names(self):
@@ -92,54 +34,11 @@ class ProfileBaseTabView(ProfileContextMixin, TabMixin, TemplateView):
         return context
 
 
-#####################
-# Profile Tab views #
-#####################
-
-
-class PartialPhotoTabMixin:
-    """Mixin to handle return partial template path for partial request."""
-
-    PHOTO_PARTIALS = {
-        "uploads": "profiles/tabs/partial/photo_partials/uploads.html",
-        "albums": "profiles/tabs/partial/photo_partials/albums.html",
-    }
-
-    def dispatch(self, request, *args, **kwargs):
-        """Set photo partial key if any. And key for partial template."""
-        self.photo_partial_key = self.get_photo_partial_key()
-        self.resolved_tab = self.photo_partial_key or self.active_tab
-        return super().dispatch(request, *args, **kwargs)
-
-    def get_photo_partial_key(self):
-        """Get photo tabs partial key in request. Return None if not found or not in allowed valid keys."""
-        query = self.request.GET.get("photo_partial_key")
-
-        return query if query in photo_tabs else None
-
-    def get_partial_rendered_string(self, context):
-        """Get partial photo tab html string."""
-        return render_to_string(
-            self.get_photo_partial_template_path(),
-            {**context, "request": self.request},
-        )
-
-    def get_photo_partial_template_path(self):
-        """Get partial photo tab template path."""
-        return self.PHOTO_PARTIALS[self.resolved_tab]
-
-
 # Base photo view.
 class ProfilePhotoBaseView(PartialPhotoTabMixin, ProfileBaseTabView):
     """Base view photo page."""
 
     profile_tab = "photos"
-
-    def get_template_names(self):
-        """Get partial template path if partial query exists, or return default full page."""
-        if self.photo_partial_key:
-            return self.get_photo_partial_template_path()
-        return super().get_template_names()
 
     def get_context_data(self, **kwargs):
         """Provide photo tabs constant for frontend."""
@@ -152,14 +51,14 @@ class ProfilePhotoBaseView(PartialPhotoTabMixin, ProfileBaseTabView):
 
 
 # Photo uploads view
-class ProfilePhotosView(ProfilePhotoBaseView):
+class PhotosUploadsView(ProfilePhotoBaseView):
     """Profile photos view that handles partial and full request."""
 
     active_tab = "uploads"
 
 
 # Photo albums view
-class ProfilePhotoAlbumFullView(ProfilePhotoBaseView):
+class PhotoAlbumFullView(ProfilePhotoBaseView):
     """View to handle profile albums page."""
 
     active_tab = "albums"
